backend/requirements.txt
# ─── Google Cloud ─────────────────────────────────────────────────────────────
google-cloud-speech>=2.21.0
google-cloud-translate>=3.15.0
google-cloud-texttospeech>=2.16.0

# ─── Google ADK ───────────────────────────────────────────────────────────────
google-adk>=0.1.0

# ─── FastAPI & WebSockets ─────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
python-multipart>=0.0.12

# ─── Config & Validation ──────────────────────────────────────────────────────
pydantic>=2.9.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# ─── Audio (local dev / testing) ─────────────────────────────────────────────
pyaudio>=0.2.14

# ─── Utilities ────────────────────────────────────────────────────────────────
tenacity>=9.0.0
